Camera2 README corrections
Fix spelling errors in Publish Camera2 README

Corrected link to README for Camera2 in this branch (previously fixed in develop)

Removed redundant “significantly”

Removed “While most applications can use the standard Camera object for its simplicity, this is an example of using Camera2.” as the original Camera object is deprecated and not recommended (despite its widespread use) per Camera2  |  Android Developers. Also, we may want to explore the use of CameraX overview  |  Android Developers as it may help our development and allow for additional features (both for customers and us).

Fixed additional grammar, spelling, and layout inconsistencies

Additional concerns:

A developer should verify some content to confirm the accuracy of old information and that descriptions are clear.

We note that the Camera API covers most live stream use cases. However, we also note that the Camera API was deprecated in favor of Camera2. Additionally, since we originally published this information, the CameraX API has been added to Android API spec. Therefore, if we recommend the Camera API over its replacements, we may cause some confusion as to what’s best to use.